+++ src/test/java/ru/vyarus/guice/ext/generator/support/bad/WrongUsageBean.java
@@ -1,0 +1,12 @@
+package ru.vyarus.guice.ext.generator.support.bad;
+
+import com.google.inject.ProvidedBy;
+import com.google.inject.internal.DynamicClassProvider;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 10.12.2014
+ */
+@ProvidedBy(DynamicClassProvider.class)
+public class WrongUsageBean {
+}