+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Update guice 3.0 -> 4.0-beta5
+* Add DynamicClassGenerator: interfaces and abstract classes may be used as guice beans with complete aop support
+* Add DynamicClassProvider for abstract classes and interfaces to use in @ProvidedBy (JIT resolution will trigger dynamic class generation )
 
 ### 1.0.1 (2014-08-16)